refactor: move virt installer backends into koan.virt subpackage
Groups the per-hypervisor create.py modules (xencreate, qcreate,
openvzcreate, vmwcreate, imagecreate) under koan/virt/ and drops the
redundant "create" suffix from each module name, since the package
location now conveys that these are VM creation backends.
