+++ koan/virt/__init__.py
@@ -1,0 +1,6 @@
+"""
+Per-hypervisor VM installer backends.
+"""
+
+# SPDX-License-Identifier: GPL-2.0-or-later
+# SPDX-FileCopyrightText: Copyright 2026 Enno Gotthold <enno.gotthold@suse.com>
