+++ koan/virt/image.py
