+++ koan/virt/qemu.py
